# openapi.yaml
openapi: 3.1.0

info:
  title: CompensationIQ API
  version: "1.0.0"
  description: |
    Salary management and compensation analytics.

    This document describes what is **implemented**, not what is planned: an OpenAPI file that
    describes endpoints nobody can call is a specification of nothing. Endpoints arrive here as
    they are built, and `docs/04-API-DESIGN.md` holds the shape of the ones that have not been.

    Two conventions run through everything below.

    **Money is an object with a string amount.** `{"amount": "1200000.00", "currency": "INR"}`,
    never a bare number: a JSON number is a double in every browser client, and a large salary
    loses its last digits on the way through one. The currency travels with the amount because
    salaries are held in the employee's own currency and this system has no default currency.

    **Collections are keyset-paged.** A page is not a numbered position in a result set; it is a
    place to resume from, given as an opaque `nextCursor`. There is no page number and no page
    count, because with this strategy there is nothing honest to put in them.

servers:
  - url: /api/v1

paths:
  /employees:
    get:
      summary: The employee directory, filtered, searched and paged
      description: |
        Returns one page of employees in a total order of `(family name, given name, id)`.

        Paging is by cursor. Ask for the first page without one, then pass the `nextCursor` from
        each response to get the next. When `nextCursor` is absent, that was the last page. The
        cursor is opaque and must not be constructed or parsed by a client: it encodes the whole
        sort key, and the sort key is free to change.
      operationId: listEmployees
      parameters:
        - name: q
          in: query
          description: |
            Free text matched anywhere within the employee's name or email address, case
            insensitively. Fragments match: `handr` finds Chandra.
          required: false
          schema: { type: string }
          example: kapoor
        - name: country
          in: query
          description: ISO-3166 alpha-2. An unknown code is a 400, not an empty page.
          required: false
          schema: { type: string, minLength: 2, maxLength: 2 }
          example: IN
        - name: department
          in: query
          required: false
          schema: { type: string }
          example: Engineering
        - name: jobTitle
          in: query
          required: false
          schema: { type: string }
          example: Software Engineer
        - name: level
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SeniorityLevel'
        - name: cursor
          in: query
          description: The `nextCursor` from the previous response. Omit for the first page.
          required: false
          schema: { type: string }
        - name: limit
          in: query
          description: |
            How many employees to return. Capped at 200 server-side: one request for the whole
            table is what pagination exists to prevent.
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: A page of employees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeePage'
        "400":
          description: |
            A filter value that cannot exist, or a cursor this API did not issue.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /employees/{id}/salary:
    put:
      summary: Change an employee's pay
      description: |
        The only way pay changes. A `PATCH` of the employee resource cannot touch salary, because
        a pay change carries a mandatory reason and must produce an audit entry - the shape of the
        API enforces what the domain enforces.

        The salary update and the audit record are written in one transaction. Either both happen
        or neither does: a salary that moved without a record of who moved it is the one outcome
        this system is built to prevent.
      operationId: changeSalary
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeSalaryRequest'
      responses:
        "200":
          description: The employee, as they now are
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        "400":
          description: |
            A field is missing, or a value cannot exist. **Also, for now, every rejection the
            domain makes**: a non-positive amount, a currency that is not the employee's country's,
            or a change to the amount already held. 04-API-DESIGN specifies 422 and 409 for those,
            and they become that when the domain distinguishes its rejections by type rather than
            by message, with the RFC 7807 work. Documented here as what the API actually returns.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        "404":
          description: No employee with that id
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        "422":
          description: The employee is TERMINATED, so their pay cannot change
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  /employees/{id}/salary-revisions:
    get:
      summary: An employee's pay history
      description: |
        Newest first, in the order the index holds them, so the read is an index scan rather than
        a sort. Append-only: there is no endpoint to edit or delete an entry, and the application's
        database role is granted neither.
      operationId: salaryRevisions
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: The log, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalaryRevision'

  /employees/filter-options:
    get:
      summary: What there is to filter the directory by
      description: |
        Derived from the people in the directory rather than configured, so a department nobody
        works in cannot be chosen and a new one appears without a deployment. Levels are the
        exception: that list belongs to the domain, not to whoever happens to be employed.
      operationId: directoryFilterOptions
      responses:
        "200":
          description: The values the filter controls should offer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryFilterOptions'

components:
  schemas:

    ChangeSalaryRequest:
      type: object
      required: [amount, currency, reason, actorId]
      properties:
        amount:
          type: string
          description: A decimal string, for the same reason salaries leave as strings.
          pattern: '^\d+(\.\d+)?$'
          example: "1380000.00"
        currency:
          type: string
          description: Must be the currency of the employee's country. Nothing is converted.
          example: INR
        reason:
          $ref: '#/components/schemas/ChangeReason'
        note:
          type: string
          description: Optional free text. Blank is stored as no note.
          example: Annual merit review
        actorId:
          type: string
          format: uuid
          description: |
            Who is making the change. **Temporary**: this belongs in the authenticated principal
            and moves there when authentication lands. It is validated against the user table
            rather than trusted, and has no default - a revision credited to an invented system
            account answers "who changed this" with a fiction.

    SalaryRevision:
      type: object
      required: [previousAmount, newAmount, reason, changedBy, changedAt]
      properties:
        previousAmount:
          $ref: '#/components/schemas/Money'
        newAmount:
          $ref: '#/components/schemas/Money'
        reason:
          $ref: '#/components/schemas/ChangeReason'
        changedBy:
          type: string
          format: uuid
        changedAt:
          type: string
          format: date-time
          description: RFC 3339 UTC. The instant the application recorded, not the instant the row was written.
        note:
          type: string

    ChangeReason:
      type: string
      enum: [MERIT, PROMOTION, MARKET_ADJUSTMENT, CORRECTION]

    DirectoryFilterOptions:
      type: object
      required: [countries, departments, jobTitles, levels]
      properties:
        countries:
          type: array
          items: { type: string, minLength: 2, maxLength: 2 }
          example: [AU, DE, GB, IN, SG, US]
        departments:
          type: array
          items: { type: string }
          example: [Customer Success, Design, Engineering, Finance, Marketing, People, Product, Sales]
        jobTitles:
          type: array
          items: { type: string }
        levels:
          type: array
          items:
            $ref: '#/components/schemas/SeniorityLevel'

    EmployeePage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        nextCursor:
          type: string
          description: |
            Where to resume. **Absent on the last page** - absent means there is no next page,
            which is why it is omitted rather than sent as null.
          example: S2Fwb29yCkFsaWNlCjExMTExMTExLTExMTEtMTExMS0xMTExLTExMTExMTExMTExMQ
        totalApprox:
          type: integer
          format: int64
          description: |
            How many employees match the filters. **Present on the first page only** - it is not
            recounted for every page, which is the cost keyset paging exists to avoid.

            **Today this value is exact.** At ten thousand employees a filtered `COUNT(*)` off
            the same indexes costs a few milliseconds, so that is what is returned - a client
            that displays it as a precise figure is displaying a precise figure.

            The name is the contract, not the current implementation: it reserves the right to
            become an estimate from `pg_class.reltuples` when the table is large enough for the
            count to hurt, at which point no client needs to change because none was ever
            promised exactness. Do not build a total that must reconcile on this field.
          example: 10000

    Employee:
      type: object
      required:
        [id, employeeNumber, givenName, familyName, email, country, department, jobTitle, level, salary]
      properties:
        id:
          type: string
          format: uuid
          description: Assigned by the application, never by the database.
        employeeNumber:
          type: string
          description: The company's own identifier for a person, distinct from `id`.
          example: E00042
        givenName: { type: string, example: Alice }
        familyName: { type: string, example: Kapoor }
        email: { type: string, format: email }
        country:
          type: string
          description: ISO-3166 alpha-2 of the market this person is employed in.
          example: IN
        department: { type: string, example: Engineering }
        jobTitle: { type: string, example: Software Engineer }
        level:
          $ref: '#/components/schemas/SeniorityLevel'
        salary:
          $ref: '#/components/schemas/Money'

    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          description: |
            A decimal **as a string**. Not a number: a JSON number is a double in a browser and a
            salary is not a value to round. Clients that do arithmetic on it should parse it into
            a decimal type, not a float.
          pattern: '^-?\d+(\.\d+)?$'
          example: "1200000.00"
        currency:
          type: string
          description: ISO-4217. The employee's own currency; nothing is converted by this API.
          minLength: 3
          maxLength: 3
          example: INR

    SeniorityLevel:
      type: string
      enum: [JUNIOR, MID, SENIOR, LEAD, PRINCIPAL]

    Problem:
      type: object
      description: RFC 7807.
      properties:
        type: { type: string, format: uri }
        title: { type: string, example: The request could not be understood }
        status: { type: integer, example: 400 }
        detail:
          type: string
          description: |
            Never contains a salary figure. Domain messages name the employee and the currencies
            involved and nothing else, because an error message is the most likely thing in the
            system to end up in a log.
          example: that is not a cursor this directory issued
        instance: { type: string, format: uri }